refactor(BP-1751): déplacement des composants résultats dans leurs composants parents
