+++ src/boites-postales/v3/resultat-bp/resultat-bp.component.ts
@@ -26,7 +26,7 @@ export class ResultatBpComponent implements OnInit {
 
   ngOnInit(): void {
     this.breakpointObserver
-      .observe([Breakpoints.Handset])
+      .observe([Breakpoints.HandsetPortrait])
       .subscribe((result) => {
         this.isMobile = result.matches;
       });
