+++ src/v3/v3.component.html
@@ -1,16 +1,21 @@
-<div class="champ-selection">
-  <div class="intitule">Je recherche :</div>
-  <mat-radio-group [(ngModel)]="choixRecherche" color="primary">
-    <mat-radio-button value="CodePostal"
-      ><div class="label">un code postal</div></mat-radio-button
-    >
-    <mat-radio-button value="BoitePostale"
-      ><div class="label">une boîte postale</div></mat-radio-button
-    >
-  </mat-radio-group>
+<div class="main-wrapper">
+  <div class="main-container">
+    <div class="champ-selection">
+      <div class="intitule">Je recherche :</div>
+      <mat-radio-group [(ngModel)]="choixRecherche" color="primary">
+        <mat-radio-button value="CodePostal"
+          ><div class="label">un code postal</div></mat-radio-button
+        >
+        <mat-radio-button value="BoitePostale"
+          ><div class="label">une boîte postale</div></mat-radio-button
+        >
+      </mat-radio-group>
+    </div>
+    @if (choixRecherche === "CodePostal") {
+    <app-codes-postaux></app-codes-postaux>
+    } @else if (choixRecherche === "BoitePostale") {
+    <app-boites-postales></app-boites-postales>
+    }
+  </div>
 </div>
-@if (choixRecherche === "CodePostal") {
-<app-codes-postaux></app-codes-postaux>
-} @else if (choixRecherche === "BoitePostale") {
-<app-boites-postales></app-boites-postales>
-}
+
