+++ src/v3/v3.component.scss
@@ -1,3 +1,15 @@
+.main-wrapper {
+display: flex;
+justify-content: center;
+}
+
+.main-container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  max-width: 1000px;
+}
+
 .champ-selection {
   margin-bottom: 15px;
 }
